+++ src/App.js
@@ -3,12 +3,12 @@ import GlobalStyle from 'GlobalStyle';
 import { Provider } from 'react-redux';
 import { store } from 'store/store';
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
-import { NavBar } from 'component';
 import { PATH } from 'constant/path';
+import NavBar from 'component/common/NavBar';
 
-import ProductList from 'page/ProductList/ProductList';
-import ShoppingCart from 'page/ShoppingCart/ShoppingCart';
-import ProductDetail from 'page/ProductDetail/ProductDetail';
+import ProductList from 'page/ProductList';
+import ShoppingCart from 'page/ShoppingCart';
+import ProductDetail from 'page/ProductDetail';
 import SnackBarContainer from 'container/SnackBarContainer';
 
 function App() {
