--- src/component/common/index.js
@@ -1,2 +1,0 @@
-export { default as Button } from 'component/common/Button/Button';
-export { default as GridLayout } from 'component/common/GridLayout/GridLayout';
