--- src/component/index.js
@@ -1,2 +1,0 @@
-export { default as NavBar } from 'component/NavBar/NavBar';
-export { default as Product } from 'component/Product/Product';
